Fix rf-filter values
rf-filter was stored as a decimal value below 1 indicating the chosen
percentage, but demod accepts integers actually - the actual percentage.

+++ librarian/forms/ondd.py
@@ -114,7 +114,7 @@ class LForm(ONDDFormBase):
         return self.preset_data.get('sample_rate', value)
 
     def preprocess_rf_filter(self, value):
-        return float(self.preset_data.get('rf_filter', value))
+        return int(self.preset_data.get('rf_filter', value))
 
     def preprocess_descrambler(self, value):
         return self.preset_data.get('descrambler', value)
